#9 Feat-[unit test component wrapper 분리]
